+++ ClickHouse.Driver.Tests/Utilities/TestUtilities.cs
@@ -277,6 +277,7 @@ public static class TestUtilities
 
         yield return new DataTypeSample("Nullable(Int32)", typeof(int?), "toInt32OrNull('123')", 123);
         yield return new DataTypeSample("Nullable(Int32)", typeof(int?), "toInt32OrNull(NULL)", DBNull.Value);
+        yield return new DataTypeSample("Nullable(String)", typeof(string), "CAST(NULL as Nullable(String))", DBNull.Value);
         yield return new DataTypeSample("Nullable(DateTime)", typeof(int?), "CAST(NULL AS Nullable(DateTime))", DBNull.Value);
 
         yield return new DataTypeSample("LowCardinality(Nullable(String))", typeof(string), "CAST(NULL AS LowCardinality(Nullable(String)))", DBNull.Value);